add sanity checks to cleanup deletion
Previously the cleanup phase just blindly tried to delete the files and
silently gave up when it couldn't find them. With this change it'll check
whether they exist and then delete them or die trying.

I suspect it might also be useful to check whether $autodir is a directory
and die if not, and to assume a .packlist is generated and die if it can't
be found, but that's up to you guys.

related to #36
related to ingydotnet/inline-pm/51